requirements.txt
openai-whisper
spacy
sounddevice
numpy
scipy
torch
coqui-tts